requirements.txt
streamlit==1.25.0
pillow
moviepy==1.0.3
numpy